+++ src/websocket_handler.cpp
@@ -2,7 +2,9 @@
 #include "led_control.h"
 #include "entity_state.h"
 #include "utils.h"
+#include "config.h"
 #include "animations.h"
+#include "secrets.h"
 
 WebSocketsClient webSocket;
 QueuedMessage queuedMessages[MAX_QUEUED_MESSAGES];
@@ -17,6 +19,92 @@ void reconnectWebSocket() {
     webSocket.disconnect();
     webSocket.begin(HA_HOST, HA_PORT, "/api/websocket");
 }
+
+
+void updateTimeAndCheckNightMode(const char* time_str) {
+    SERIAL_PRINTF("Received time update: %s\n", time_str);
+
+    if (!time_str || strlen(time_str) < 5) {
+        SERIAL_PRINTLN("Invalid time string received");
+        return;
+    }
+
+    int hour = 0, minute = 0;
+    if (sscanf(time_str, "%d:%d", &hour, &minute) != 2) {
+        SERIAL_PRINTF("Failed to parse time string: %s\n", time_str);
+        return;
+    }
+
+    if (hour < 0 || hour > 23 || minute < 0 || minute > 59) {
+        SERIAL_PRINTF("Invalid time values: %02d:%02d\n", hour, minute);
+        return;
+    }
+
+    bool newIsNightMode;
+    if (NIGHT_START_HOUR > NIGHT_END_HOUR) {
+        // Night mode spans midnight
+        newIsNightMode = (hour >= NIGHT_START_HOUR || hour < NIGHT_END_HOUR);
+    } else {
+        // Night mode doesn't span midnight
+        newIsNightMode = (hour >= NIGHT_START_HOUR && hour < NIGHT_END_HOUR);
+    }
+    if (newIsNightMode != isNightMode) {
+        isNightMode = newIsNightMode;
+        SERIAL_PRINTF("Night mode changed to: %s (Time: %02d:%02d)\n", isNightMode ? "ON" : "OFF", hour, minute);
+        for (int y = 0; y < ROWS; y++) {
+            for (int x = 0; x < COLS; x++) {
+                updateLED(x, y);
+            }
+        }
+    } else {
+        SERIAL_PRINTF("Night mode unchanged: %s (Time: %02d:%02d)\n", isNightMode ? "ON" : "OFF", hour, minute);
+    }
+}
+
+
+void toggleEntity(int x, int y) {
+    for (int i = 0; i < NUM_MAPPINGS; i++) {
+        if (entityMappings[i].x == x && entityMappings[i].y == y) {
+            DynamicJsonDocument doc(1024);
+            doc["id"] = messageId++;
+            doc["type"] = "call_service";
+            
+            if (isMediaPlayer(entityMappings[i].entity_id)) {
+                doc["domain"] = "media_player";
+                doc["service"] = "media_play_pause";
+                SERIAL_PRINTF("Attempting to play/pause media player: %s\n", entityMappings[i].entity_id);
+            } else if (isLight(entityMappings[i].entity_id)) {
+                doc["domain"] = "light";
+                doc["service"] = "toggle";
+                SERIAL_PRINTF("Attempting to toggle light: %s\n", entityMappings[i].entity_id);
+            } else if (isSwitch(entityMappings[i].entity_id)) {
+                doc["domain"] = "homeassistant";
+                doc["service"] = "toggle";
+                SERIAL_PRINTF("Attempting to toggle switch: %s\n", entityMappings[i].entity_id);
+            } else {
+                SERIAL_PRINTF("Unknown entity type: %s\n", entityMappings[i].entity_id);
+                return;
+            }
+            
+            doc["target"]["entity_id"] = entityMappings[i].entity_id;
+
+            String message;
+            serializeJson(doc, message);
+            SERIAL_PRINTF("Sending message: %s\n", message.c_str());
+
+            bool sent = webSocket.sendTXT(message);
+            if (sent) {
+                SERIAL_PRINTF("Message sent successfully for entity at (%d, %d): %s\n", x, y, entityMappings[i].entity_id);
+            } else {
+                SERIAL_PRINTF("Failed to send message for entity at (%d, %d): %s\n", x, y, entityMappings[i].entity_id);
+            }
+
+            return;
+        }
+    }
+    SERIAL_PRINTF("No entity found at (%d, %d) to toggle\n", x, y);
+}
+
 
 void webSocketEvent(WStype_t type, uint8_t * payload, size_t length) {
     SERIAL_PRINTF("WebSocket event type: %d\n", type);
